Refactor freetype
