+++ src/meson.build
@@ -66,28 +66,7 @@ subdir('font')
 
 subdir('mixer')
 
-if sdl_api != 3
-
-if freetype_dep.found()
-    _freetype = py.extension_module(
-        '_freetype',
-        [
-            'freetype/ft_cache.c',
-            'freetype/ft_wrap.c',
-            'freetype/ft_render.c',
-            'freetype/ft_render_cb.c',
-            'freetype/ft_layout.c',
-            'freetype/ft_unicode.c',
-            '_freetype.c',
-        ],
-        c_args: warnings_error + warnings_temp_freetype,
-        dependencies: pg_base_deps + freetype_dep,
-        install: true,
-        subdir: pg,
-    )
-endif
-
-endif
+subdir('freetype')
 
 if portmidi_dep.found()
     pypm = py.extension_module(
